+++ prefs.h
@@ -7,17 +7,17 @@ class Prefs
 {
 public:
     enum _modes { _PHASH, _SSIM };
-    short _ComparisonMode = _PHASH;
-    ushort _thumbnails = thumb4;
+    int _ComparisonMode = _PHASH;
+    int _thumbnails = thumb4;
 
-    short _thresholdPhash = 7;
-    short _thresholdPhashOriginal = 7;
+    int _thresholdPhash = 7;
+    int _thresholdPhashOriginal = 7;
 
     double _thresholdSSIM = 0.90;
-    short _ssimBlockSize = 16;
+    int _ssimBlockSize = 16;
 
-    short _differentDurationModifier = 4;
-    short _sameDurationModifier = 1;
+    int _differentDurationModifier = 4;
+    int _sameDurationModifier = 1;
 };
 
 #endif // PREFS_H
